requirements.txt
django
pillow
django-ajaximage
